+++ src/Utils/PoissonDisc.cpp
@@ -14,17 +14,17 @@ PoissonDisc::PoissonDisc()
 	seed(rd());
 }
 
-PoissonDisc::PoissonDisc(int seed_)
+PoissonDisc::PoissonDisc(unsigned seed_)
 {
 	seed(seed_);
 }
 
-void PoissonDisc::seed(int seed)
+void PoissonDisc::seed(unsigned seed)
 {
 	mt.seed(seed);
 }
 
-void PoissonDisc::generate2D(size_t width, size_t height, double minDistance, size_t iterationLimit)
+void PoissonDisc::generate2D(size_t width, size_t height, double minDistance, unsigned iterationLimit)
 {
 	std::uniform_real_distribution<double> random(0.0, 1.0);
 
@@ -53,7 +53,7 @@ void PoissonDisc::generate2D(size_t width, size_t height, double minDistance, si
 	{
 		Vector2 origin = getNextActivePoint2D();
 
-		for (size_t i = 0; i < iterationLimit; ++i)
+		for (unsigned i = 0; i < iterationLimit; ++i)
 		{
 			Vector2 point = generateNewPoint2D(origin, minDistance);
